requirements.txt
asgiref==3.5.0
boto3==1.21.13
botocore==1.24.13
certifi==2021.10.8
charset-normalizer==2.0.12
Cython==0.29.28
dj-database-url==0.5.0
Django==4.0.3
django-heroku==0.3.1
djangorestframework==3.13.1
gluonnlp==0.10.0
graphviz==0.8.4
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
JPype1==1.3.0
lxml==4.8.0
mxnet==1.9.0
numpy==1.22.2
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
s3transfer==0.5.2
sentencepiece==0.1.96
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==6.0.0
